# Conflicts:
#	Travely_Manager/app/src/main/java/com/hyeran/android/travely_manager/ReserveDetailFragment.kt
#	Travely_Manager/app/src/main/java/com/hyeran/android/travely_manager/network/NetworkService.kt
#	Travely_Manager/app/src/main/res/layout/fragment_reserve_detail.xml

+++ Travely_Manager/app/src/main/java/com/hyeran/android/travely_manager/network/NetworkService.kt
@@ -51,8 +51,8 @@ interface NetworkService {
     @PUT("/api/owner/reserve/{reserveIdx}")
     fun putStorePickUpResponse(
             @Header("jwt") jwt: String?,
-            @Path("reserveIdx") reserveIdx: Int
-    ): Call<Any>
+            @Path("reserveIdx") reserveIdx: Long
+    ) : Call<Any>
 
     // 예약코드 조회
     @GET("/api/owner/reserve/{storeIdx}/{reserveCode}")
